requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.9.1
click==6.7
cryptography==1.8.1
Flask==0.12.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
keyring==10.3.2
MarkupSafe==1.0
packaging==16.8
pycparser==2.17
pyparsing==2.2.0
python-vlc==1.1.2
pywin32-ctypes==0.0.1
six==1.10.0
Werkzeug==0.12.1
yagmail==0.6.161
